River and RiverSide classes
RiverSide represents either the left, or the right "state" of a river.

River has two RiverSides. Left and Right. On each creation of River, it checks whether the RiverSides are valid. 
River has a method "Cross". Cross takes the number of Missionaries and Cannibals to move. It knows where the boat is so it isn't needed as an input. Then Cross creates a new River, which validates the states and the Cross and returns it.

The idea is we make a BST, and add each state of River, and then we Cross and add the output of that. 

We can figure out if any River is in a valid state by calling the method "isValid". If it is we continue down the BST, otherwise we remove it and go back to the prior state.